startbooks/requirements.txt
asgiref==3.8.1
channels==4.2.0
Django==5.1.6
djangorestframework==3.15.2
sqlparse==0.5.3
typing_extensions==4.12.2
